Replace placeholder component in NamespaceSelectFilter
