+++ src/renderer/components/+namespaces/namespace-select.tsx
@@ -7,8 +7,6 @@ import { Select, SelectOption, SelectProps } from "../select";
 import { cssNames } from "../../utils";
 import { Icon } from "../icon";
 import { namespaceStore } from "./namespace.store";
-import { FilterIcon } from "../item-object-list/filter-icon";
-import { FilterType } from "../item-object-list/page-filters.store";
 import { kubeWatchApi } from "../../api/kube-watch-api";
 
 interface Props extends SelectProps {
@@ -35,7 +33,7 @@ export class NamespaceSelect extends React.Component<Props> {
     ]);
   }
 
-  @computed get options(): SelectOption[] {
+  @computed.struct get options(): SelectOption[] {
     const { customizeOptions, showClusterOption, showAllNamespacesOption } = this.props;
     let options: SelectOption[] = namespaceStore.items.map(ns => ({ value: ns.getName() }));
 
@@ -79,58 +77,3 @@ export class NamespaceSelect extends React.Component<Props> {
   }
 }
 
-@observer
-export class NamespaceSelectFilter extends React.Component {
-  @computed get placeholder(): React.ReactNode {
-    const namespaces = namespaceStore.contextNamespaces;
-
-    switch (namespaces.length) {
-      case 0:
-      case namespaceStore.allowedNamespaces.length:
-        return <>All namespaces</>;
-      case 1:
-        return <>Namespace: {namespaces[0]}</>;
-      default:
-        return <>Namespaces: {namespaces.join(", ")}</>;
-    }
-  }
-
-  formatOptionLabel = ({ value: namespace, label }: SelectOption) => {
-    if (namespace) {
-      const isSelected = namespaceStore.hasContext(namespace);
-
-      return (
-        <div className="flex gaps align-center">
-          <FilterIcon type={FilterType.NAMESPACE}/>
-          <span>{namespace}</span>
-          {isSelected && <Icon small material="check" className="box right"/>}
-        </div>
-      );
-    }
-
-    return label;
-  };
-
-  onChange = ([{ value: namespace }]: SelectOption[]) => {
-    if (namespace) {
-      namespaceStore.toggleContext(namespace);
-    } else {
-      namespaceStore.resetContext(); // "All namespaces" clicked, empty list considered as "all"
-    }
-  };
-
-  render() {
-    return (
-      <NamespaceSelect
-        isMulti={true}
-        showAllNamespacesOption={true}
-        closeMenuOnSelect={false}
-        isOptionSelected={() => false}
-        controlShouldRenderValue={false}
-        placeholder={this.placeholder}
-        onChange={this.onChange}
-        formatOptionLabel={this.formatOptionLabel}
-      />
-    );
-  }
-}
